+++ src/components/BitcoinCharts/BitcoinCharts.js
