+++ src/components/BitcoinCharts/styles.js
