+++ src/components/CoinPageChart/CoinPageChart.js
