+++ src/components/CoinPageChart/styles.js
