+++ src/components/CoinsTable/styles.js
